Hero turret controller updates

+++ PolySTAR-Taproot-project/src/algorithms/gravity_feed_forward.cpp
@@ -10,7 +10,7 @@ namespace algorithms
 {
 
 GravityFeedForward::GravityFeedForward(const FeedForwardConfig &ffConfig, float x, float y): 
-    offsetAngle(atan(y/x)), 
+    offsetAngle(atanf(y/x)), 
     FeedForward(ffConfig) 
 {}
